+++ src/test/commentTestSamples/issue-111-throw-finally.py
@@ -1,0 +1,7 @@
+# nodes: 4,
+# exits: 1
+def raise_and_finally():
+    try:
+        raise RuntimeError()
+    finally:
+        print("Oh no!")