+++ src/test/commentTestSamples/issue-111-throw-finally.ts
@@ -1,0 +1,11 @@
+/*
+nodes: 4,
+exits: 1
+ */
+function throwAndFinally() {
+  try {
+    throw new Error()
+  } finally {
+    console.log("Oh no!")
+  }
+}